Task Completed
Create a new plugin with 2 forms. The plugin is only available for logged in users, no access for guest users.

Form 1 will accept the following inputs: First name, last name, username, email, password and once it is submitted it will create a new user. The user created can log in using their username and password.

Form 2 will list all the courses assigned to the company (dropdown) and all the active users in the company (dropdown), course score and course completion date. Once the form is submitted it will records a course completion for the user with the selected course, selected user, completion date and score.